fix: Pass through double dollar signs($$)
Dollar signs are given special treatment in String.prototype.replace()
when used in the replacement string(second parameter). To avoid this
behavior, a function is used to provide the replacement string, as the
special rules applying to dollar signs does not apply to the function
return value.

This fixes a bug where the special svelte variables $$slots and $$props
would not work when this plugin is used, as $$ would be replaced with a
single $ in the compiled code.